Corrected image vectorization script

+++ Image vectorization/tiff_image_vectorization.h
@@ -3,16 +3,16 @@
 #include <tiffio.h>
 
 typedef struct {
-    uint32_t width;
-    uint32_t height;
-    uint16_t channels;
+    int width;
+    int height;
+    int channels;
     float* data;
 } ImageTensor;
 
 int read_tiff_info(const char* filename,
-                   uint32_t* width,
-                   uint32_t* height,
-                   uint16_t* channels);
+                   int* width,
+                   int* height,
+                   int* channels);
 
 ImageTensor* load_tiff_as_tensor(const char* filename);
